netcdf: fix wrapping of datapoint extractor
A bug existed when extracting datapoints close to the LHS of 180deg,
because this was the wrapping position of the dataset. It worked for
values on the RHS of the line because negative indexes in Python pull
from the right hand side of the dataset, whereas indexes > len(data)
generate an IndexError. To handle this correctly, check whether the
dataset wraps, and then use modulo arithmatic.

We can move the check for whether or not the dataset wraps to be outside
the tight loop.

Fixes: http://tuscany/redmine/issues/93