Only import required module; only generate JSON in portal.py
If one module has a syntax error, the whole portal would blow up because
absolutely everything would get imported. With this change, the portal
will only import the required module for the dataset (which assumes that
all datasets are laid out the same), and thus a syntax error will only
break one dataset.

Additionally we want to add things to the JSON return in the top level
(diagnostics, version number, etc.) so we should return the dict from
process() and output to JSON in portal.py.