compmap: relayout panels to handle small screen space
This moves the outputs panel to be on the RHS and renames it to Outputs,
since it also includes data downloads, climatology, etc. This also
leaves us with more space to do things like break the layers dialog out
of the map (which confuses people, or they don't discover it).