compmap: fix date weirdness
If we had the 31st day of the month set, and we set a month with less
days in monthly, we are accidentally advanced to the next month.
Similarly this can come up with leap years and February. Basically,
SURPRISE!

Fix this by making monthly/yearly/etc. set the date as the 1st.

Fixes: http://tuscany/redmine/issues/149