Fix extents handling
Zoom to bounds, not to centre at fixed zoom level. Needs lots more
testing. Some of the regions in the region config need tweaking.

Build restricted extents from the union of all the regions extents,
instead of specifying it in the config file. This avoids problems like
we had where Norfolk Is wasn't within the restricted extents. Means we
need to add the low resolution landmask in for areas outside the South
Pacific.