Change server configs to be class based instead of dictionary based
This allows for server configs to be inherited from each other and from
a default config, e.g. hoapp2 inherits from www4 inherits from default.
It also allows for attributes to be defined as methods using @property.

Code to glue the old config dict to the new class/attributes format is
in BaseConfig (which default inherits).